+++ Vk_RubiksCube/structs/Vk_RenderData.h
@@ -16,7 +16,7 @@ struct RenderData
     std::vector<Vertex> vertices;
     std::vector<uint32_t> indices;
 
-    std::unordered_map<uint32_t, std::pair<size_t, size_t>> material_index_ranges;
+    std::unordered_map<uint32_t, std::pair<uint32_t, uint32_t>> material_index_ranges;
 };
 
 #endif
